Made the unit dynamic

+++ backend/app/models/stock_history.py
@@ -1,15 +1,15 @@
 import enum
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, Integer, String, DateTime, Enum, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, Enum, ForeignKey
 
 from app.database import Base
 
 
 class StockChangeReason(str, enum.Enum):
-    ADJUSTMENT = "adjustment"  # owner manually added/removed stock
-    SALE = "sale"  # decremented because a customer placed an order
-    CANCELLED = "cancelled"  # returned to shelf because an order was cancelled
+    ADJUSTMENT = "adjustment"
+    SALE = "sale"
+    CANCELLED = "cancelled"
 
 
 class StockHistory(Base):
@@ -17,16 +17,14 @@ class StockHistory(Base):
 
     id = Column(Integer, primary_key=True, index=True)
 
-    # Nullable + a name snapshot, same pattern as OrderItem: history should
-    # still read correctly even after the product itself is deleted.
     product_id = Column(Integer, ForeignKey("products.id"), nullable=True, index=True)
     product_name = Column(String(150), nullable=False)
 
     owner_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
 
-    change = Column(Integer, nullable=False)  # signed: +10 restock, -2 sale
+    change = Column(Numeric(12, 3), nullable=False)
     reason = Column(Enum(StockChangeReason), nullable=False)
-    previous_stock = Column(Integer, nullable=False)
-    new_stock = Column(Integer, nullable=False)
+    previous_stock = Column(Numeric(12, 3), nullable=False)
+    new_stock = Column(Numeric(12, 3), nullable=False)
 
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
