+++ backend/app/schemas/stock.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -7,7 +8,10 @@ from app.models.stock_history import StockChangeReason
 
 
 class StockAdjustment(BaseModel):
-    delta: int = Field(description="Positive to add stock, negative to remove it.")
+    delta: Decimal = Field(
+        decimal_places=3,
+        description="Positive to add stock, negative to remove it.",
+    )
 
 
 class StockHistoryOut(BaseModel):
@@ -16,8 +20,8 @@ class StockHistoryOut(BaseModel):
     id: int
     product_id: Optional[int] = None
     product_name: str
-    change: int
+    change: Decimal
     reason: StockChangeReason
-    previous_stock: int
-    new_stock: int
+    previous_stock: Decimal
+    new_stock: Decimal
     created_at: datetime
